Viewの変更

+++ backend/resources/views/user/roads/index.blade.php
@@ -26,14 +26,18 @@
                                                     <p class="leading-relaxed mb-3">{{ Str::limit($road->description, 100, ' ...続きを読む') }}</p>
                                                     <div class="flex justify-between">
                                                         <div class="flex">
-                                                            <button  onclick="location.href='{{ route('user.roads.edit', ['road' => $road->id ]) }}'" class="text-black bg-yellow-400 border-0 py-2 px-4 mr-2 focus:outline-none hover:bg-yellow-500 rounded">編集</button>
-                                                            <form id="delete_{{ $road->id }}" method="post" action="{{ route('user.roads.destroy', ['road' => $road->id ]) }}">
-                                                                @csrf
-                                                                @method('delete')
-                                                                <div class="w-full flex justify-around">
-                                                                    <a href="#" data-id="{{ $road->id }}" onclick="deletePost(this)" class="text-white bg-red-400 border-0 py-2 px-4 focus:outline-none hover:bg-red-500 rounded">削除</a>
-                                                                </div>
-                                                            </form>
+                                                            @can('update', $road)
+                                                                <button  onclick="location.href='{{ route('user.roads.edit', ['road' => $road->id ]) }}'" class="text-black bg-yellow-400 border-0 py-2 px-4 mr-2 focus:outline-none hover:bg-yellow-500 rounded">編集</button>
+                                                            @endcan
+                                                            @can('delete', $road)
+                                                                <form id="delete_{{ $road->id }}" method="post" action="{{ route('user.roads.destroy', ['road' => $road->id ]) }}">
+                                                                    @csrf
+                                                                    @method('delete')
+                                                                    <div class="w-full flex justify-around">
+                                                                        <a href="#" data-id="{{ $road->id }}" onclick="deletePost(this)" class="text-white bg-red-400 border-0 py-2 px-4 focus:outline-none hover:bg-red-500 rounded">削除</a>
+                                                                    </div>
+                                                                </form>
+                                                            @endcan
                                                         </div>
                                                         <div>
                                                             <p class="leading-relaxed text-right">{{ $road->created_at->format('Y-m-d') }}</p>
